requirements.txt
Flask
requests
beautifulsoup4
google-cloud-aiplatform
google-auth
